Rename opTooltipContainer to opTooltipService
